v0.1.7: Startup script support and server.go code split
Split cmd/server/server.go (861 lines) into 6 focused files:
- server.go: Cmd, init, shared helpers
- list.go: list, show, rename
- create.go: create, interactive pickers, volume helpers, user_data
- lifecycle.go: delete, start, stop, reboot, resize, rebuild, console
- metadata.go: ips, metadata
- volume_attach.go: attach-volume, detach-volume

Add startup script (user_data) support to server create:
- --user-data <file>: read file, validate size, base64-encode
- --user-data-raw <string>: encode inline string
- --user-data-url <url>: wrap as #include directive
- Mutual exclusion between the three flags
- 16 KiB size limit validation
- Warning for Windows flavors

Add docs/startup-script.md usage guide and update README.
