+++ README.md
@@ -78,7 +78,7 @@ conoha server rename <server-id-or-name> new-name
 | コマンド | 説明 |
 |---------|------|
 | `conoha auth` | 認証管理（login / logout / status / list / switch / token / remove） |
-| `conoha server` | サーバー管理（list / show / create / delete / start / stop / reboot / resize / rebuild / rename / console） |
+| `conoha server` | サーバー管理（list / show / create / delete / start / stop / reboot / resize / rebuild / rename / console / ips / metadata） |
 | `conoha flavor` | フレーバー一覧・詳細（list / show） |
 | `conoha keypair` | SSH キーペア管理（list / create / delete） |
 | `conoha volume` | ブロックストレージ管理（list / show / create / delete / types / backup） |
@@ -89,6 +89,42 @@ conoha server rename <server-id-or-name> new-name
 | `conoha storage` | オブジェクトストレージ（container / ls / cp / rm / publish） |
 | `conoha identity` | アイデンティティ管理（credential / subuser / role） |
 | `conoha config` | CLI 設定管理（show / set / path） |
+
+## サーバー作成
+
+対話的にサーバーを作成できます（フレーバー、イメージ、キーペアを選択）：
+
+```bash
+conoha server create --name my-server
+```
+
+フラグで直接指定も可能：
+
+```bash
+conoha server create --name my-server \
+  --flavor g2l-t-c2m1 \
+  --image <image-id> \
+  --key-name my-key \
+  --admin-pass 'P@ssw0rd'
+```
+
+### スタートアップスクリプト
+
+サーバー作成時に初期設定スクリプトを指定できます：
+
+```bash
+# ファイルから
+conoha server create --name my-server --user-data ./init.sh
+
+# インライン
+conoha server create --name my-server --user-data-raw '#!/bin/bash
+apt update && apt install -y nginx'
+
+# URL 指定
+conoha server create --name my-server --user-data-url https://example.com/setup.sh
+```
+
+詳細は [docs/startup-script.md](docs/startup-script.md) を参照してください。
 
 ## 設定
