+++ cmd/server/list.go
@@ -1,0 +1,144 @@
+package server
+
+import (
+	"fmt"
+	"os"
+	"time"
+
+	"github.com/spf13/cobra"
+
+	"github.com/crowdy/conoha-cli/cmd/cmdutil"
+	"github.com/crowdy/conoha-cli/internal/model"
+	"github.com/crowdy/conoha-cli/internal/output"
+)
+
+var listCmd = &cobra.Command{
+	Use:   "list",
+	Short: "List all servers",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		compute, err := getComputeAPI(cmd)
+		if err != nil {
+			return err
+		}
+		servers, err := compute.ListServers()
+		if err != nil {
+			return err
+		}
+
+		flavors, err := compute.ListFlavors()
+		if err != nil {
+			return err
+		}
+		flavorMap := make(map[string]string, len(flavors))
+		for _, f := range flavors {
+			flavorMap[f.ID] = f.Name
+		}
+
+		type serverRow struct {
+			ID     string `json:"id"`
+			Name   string `json:"name"`
+			Status string `json:"status"`
+			Flavor string `json:"flavor"`
+			Tag    string `json:"tag"`
+		}
+		rows := make([]serverRow, len(servers))
+		for i, s := range servers {
+			flavorName := flavorMap[s.Flavor.ID]
+			if flavorName == "" {
+				flavorName = s.Flavor.ID
+			}
+			rows[i] = serverRow{
+				ID:     s.ID,
+				Name:   s.Name,
+				Status: s.Status,
+				Flavor: flavorName,
+				Tag:    s.Metadata["instance_name_tag"],
+			}
+		}
+
+		return output.New(cmdutil.GetFormat(cmd)).Format(os.Stdout, rows)
+	},
+}
+
+var showCmd = &cobra.Command{
+	Use:   "show <id|name>",
+	Short: "Show server details",
+	Args:  cmdutil.ExactArgs(1),
+	RunE: func(cmd *cobra.Command, args []string) error {
+		compute, err := getComputeAPI(cmd)
+		if err != nil {
+			return err
+		}
+		server, err := compute.FindServer(args[0])
+		if err != nil {
+			return err
+		}
+
+		format := cmdutil.GetFormat(cmd)
+		if format != "" && format != "table" {
+			return output.New(format).Format(os.Stdout, server)
+		}
+
+		// Human-readable key-value output
+		printServerDetail(server)
+		return nil
+	},
+}
+
+func printServerDetail(s *model.Server) {
+	jst := time.FixedZone("JST", 9*60*60)
+
+	fmt.Printf("ID:        %s\n", s.ID)
+	fmt.Printf("Name:      %s\n", s.Name)
+	fmt.Printf("Status:    %s\n", s.Status)
+	fmt.Printf("Flavor:    %s\n", s.Flavor.ID)
+	fmt.Printf("Image:     %s\n", s.ImageID)
+	fmt.Printf("Key Name:  %s\n", s.KeyName)
+	fmt.Printf("Tenant:    %s\n", s.TenantID)
+	fmt.Printf("Created:   %s (%s JST)\n",
+		s.Created.Format(time.RFC3339),
+		s.Created.In(jst).Format("2006-01-02 15:04"))
+	if !s.Updated.IsZero() {
+		fmt.Printf("Updated:   %s (%s JST)\n",
+			s.Updated.Format(time.RFC3339),
+			s.Updated.In(jst).Format("2006-01-02 15:04"))
+	}
+
+	if len(s.Addresses) > 0 {
+		fmt.Println("Addresses:")
+		for net, addrs := range s.Addresses {
+			for _, a := range addrs {
+				fmt.Printf("  %s: %s (v%d, %s)\n", net, a.Addr, a.Version, a.Type)
+			}
+		}
+	}
+
+	if len(s.Metadata) > 0 {
+		fmt.Println("Metadata:")
+		for k, v := range s.Metadata {
+			fmt.Printf("  %s: %s\n", k, v)
+		}
+	}
+}
+
+var renameCmd = &cobra.Command{
+	Use:   "rename <id|name> <new-name>",
+	Short: "Rename a server",
+	Args:  cmdutil.ExactArgs(2),
+	RunE: func(cmd *cobra.Command, args []string) error {
+		compute, err := getComputeAPI(cmd)
+		if err != nil {
+			return err
+		}
+		server, err := compute.FindServer(args[0])
+		if err != nil {
+			return err
+		}
+		renamed, err := compute.RenameServer(server.ID, args[1])
+		if err != nil {
+			return err
+		}
+		fmt.Fprintf(os.Stderr, "Server renamed: %s -> %s\n", args[0], renamed.Name)
+		return nil
+	},
+}
